// src/app/pages/journal/data-access/journal.interface.ts
import { ObjId } from '@shared/interfaces/common.interface';
import { JournalType } from './journal.dto';

export interface Journal {
  id: ObjId | null;
  type: JournalType;
  journalDate: string;
  journalEndDate: string;
  publishDate: string | null;
  status: string;
  acknowledgementComment: string | null;
  updatedAt: string | null;
  createdAt: string | null;
  semesterId: number | null;
  semesterName: string | null;
  classId: number;
  classDisplayName: string;
  levelId: number;
  levelDisplayName: string;
  roleId: number;
  roleDisplayName: string;
  guardianId: number | null;
  studentId: number;
  displayName: string;
  userId: number;
  nationalId: string;
  displayPhoneNumber: string;
  gender: string;
  dateOfBirth: string;
  pioneerId: string;
  registrationDate: string;
  viewedByGuardian: boolean | null;
  arabicTeacherDisplayName: string | null;
  englishTeacherDisplayName: string | null;
  scienceTeacherDisplayName: string | null;
  healthAndCareTeacherDisplayName: string | null;
  islamicTeacherDisplayName: string | null;
  arabicTeacherNote: string | null;
  scienceTeacherNote: string | null;
  englishTeacherNote: string | null;
  healthAndCareTeacherNote: string | null;
  islamicTeacherNote: string | null;
  taskDone: string | null;
  meal: string[];
  lunch: string[];
  snack: string[];
  skillClubs: string[];
  guardianDisplayName: string;
  needGuardianAttention: boolean;
  healthAndCare: string | null;
}